Add basic support for markers in requirements.txt
Basically this allows liccheck to work on requirements.txt file having
lines such as:
  python-openid;python_version<="2.7"
  python3-openid;python_version>="3.0"

Before, it would try to evaluate both lines in both python2 and python3
resulting in liccheck crashing in both environment.
With this commit, it calls to pkg_resources.evaluate_marker and if that
function returns false, it ignores the requirement.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>