fix: Include `tox` in test-requirements.txt
The contribution instructions direct new code contributors to run test using
`tox`. However, `tox` is not included in the `test-requirements.txt` file,
requiring an extra step for new contributors to install `tox` manually.

This commit adds `tox` to the `test-requirements.txt` file so that new code
contributors automatically get it when they install the test requirements.